* master: (21 commits)
  ABB TACW fw ver 1.6.5 adaptions (#4626)
  heidelberg: remove fw register diag (#5125)
  chore: improve issue template (#5116)
  Fix charger rater/timer logic for vehicle reconnect
  Enyaq: add soc limit
  eTron: fix api error
  Go-E: fix session energy reset on disconnect (BC) (#5119)
  FritzDECT: increase session timeout
  Gokrazy: add sqlite to /perm
  increase community data update interval (5s > 10s)
  chore: improve proxy example by adding rtu
  Docker: expose mDNS (#5093)
  Sessions: add odometer (#5069)
  Docker: don't force master branch when releasing to enable bugfix releases
  chore: fix comment
  Fix race condition in loading templates
  Fix a race condition when stopping the charging session
  Show version in api
  chore: add some docs
  Publish savings data on startup (#5048)
  ...

+++ vehicle/vw/id/types.go
@@ -53,7 +53,7 @@ type ChargingStatus struct {
 	ChargeMode                         string  // invalid
 	RemainingChargingTimeToCompleteMin int     `json:"remainingChargingTimeToComplete_min"`
 	ChargePowerKW                      float64 `json:"chargePower_kW"`
-	ChargeRateKmph                     int     `json:"chargeRate_kmph"`
+	ChargeRateKmph                     float64 `json:"chargeRate_kmph"`
 }
 
 // ChargingSettings is the /status.chargingSettings api
